# resources/swagger.json
{"openapi":"3.0.0","paths":{"/v1/auth/signup":{"post":{"operationId":"AuthController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Auth"]}},"/v1/auth/refresh":{"get":{"operationId":"AuthController_refresh","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Auth"]}},"/v1/auth/signin":{"post":{"operationId":"AuthController_signin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Auth"]}},"/v1/auth/email/forgot-password/{email}":{"get":{"operationId":"AuthController_sendEmailForgotPassword","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Auth"]}},"/v1/auth/email/reset-password":{"post":{"operationId":"AuthController_setNewPassord","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Auth"]}},"/v1/user/me":{"get":{"operationId":"UserController_getMe","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["Users"]}},"/v1/user/exists/{id}":{"get":{"operationId":"UserController_exists","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Users"]}},"/v1/user":{"patch":{"operationId":"UserController_editUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["Users"]}},"/v1/user/role":{"patch":{"operationId":"UserController_changeUserRole","parameters":[{"name":"role","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["Users"]}},"/v1/camps":{"get":{"operationId":"CampController_getCamps","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CampWithLocationsEntity"}}}}}},"tags":["Camps"]},"post":{"operationId":"CampController_createCamp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampEntity"}}}}},"tags":["Camps"]}},"/v1/camps/{campId}":{"get":{"operationId":"CampController_getCampById","parameters":[{"name":"campId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Camps"]},"patch":{"operationId":"CampController_editCampById","parameters":[{"name":"campId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampEntity"}}}}},"tags":["Camps"]},"delete":{"operationId":"CampController_deleteCampById","parameters":[{"name":"campId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Camps"]}},"/v1/camps/{campName}":{"get":{"operationId":"CampController_getCampByName","parameters":[{"name":"campName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Camps"]}},"/v1/camps/{campId}/location":{"post":{"operationId":"CampController_addCampLocationByCampId","parameters":[{"name":"campId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Camps"]}},"/v1/camps/brc_data/{year}":{"get":{"operationId":"CampController_getRemoteCamps","parameters":[{"name":"year","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Camps"]}},"/v1/camps/mock_data/{count}":{"get":{"operationId":"CampController_createMockCamps","parameters":[{"name":"count","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Camps"]}},"/v1/arts":{"get":{"operationId":"ArtController_getArts","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ArtWithLocationsEntity"}}}}}},"tags":["Arts"]},"post":{"operationId":"ArtController_createArt","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArtDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArtEntity"}}}}},"tags":["Arts"]}},"/v1/arts/{artId}":{"get":{"operationId":"ArtController_getArtById","parameters":[{"name":"artId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Arts"]},"patch":{"operationId":"ArtController_editArtById","parameters":[{"name":"artId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArtDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArtEntity"}}}}},"tags":["Arts"]},"delete":{"operationId":"ArtController_deleteArtById","parameters":[{"name":"artId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Arts"]}},"/v1/arts/{artName}":{"get":{"operationId":"ArtController_getArtByName","parameters":[{"name":"artName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Arts"]}},"/v1/arts/{artId}/location":{"post":{"operationId":"ArtController_addArtLocationByArtId","parameters":[{"name":"artId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Arts"]}},"/v1/arts/brc_data/{year}":{"get":{"operationId":"ArtController_getRemoteArts","parameters":[{"name":"year","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Arts"]}},"/v1/arts/mock_data/{count}":{"get":{"operationId":"ArtController_createMockArts","parameters":[{"name":"count","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Arts"]}}},"info":{"title":"Unofficial BRC Backend API","description":"Unofficial BRC Backend API","version":"1.0","contact":{}},"tags":[{"name":"NestJs","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"AuthDto":{"type":"object","properties":{"email":{"type":"string","default":"email@email.com"},"password":{"type":"string"},"phone":{"type":"string"},"token":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"}},"required":["email","password","phone"]},"ResetPasswordDto":{"type":"object","properties":{"email":{"type":"string"},"newPassword":{"type":"string"},"newPasswordToken":{"type":"string"},"currentPassword":{"type":"string"}},"required":["email","newPassword","newPasswordToken","currentPassword"]},"UserEntity":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"},"firstName":{"type":"string","nullable":true},"lastName":{"type":"string","nullable":true}},"required":["id","email","role","firstName","lastName"]},"UserDto":{"type":"object","properties":{"email":{"type":"string","default":"email@email.com"},"username":{"type":"string"},"name":{"type":"string"}},"required":["email"]},"CampWithLocationsEntity":{"type":"object","properties":{"uid":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"year":{"type":"number"},"name":{"type":"string"},"url":{"type":"string","nullable":true},"contact_email":{"type":"string","nullable":true},"hometown":{"type":"string","nullable":true},"description":{"type":"string","nullable":true},"locations":{"type":"array","items":{"type":"object"}}},"required":["uid","createdAt","updatedAt","year","name","url","contact_email","hometown","description","locations"]},"CampDto":{"type":"object","properties":{"name":{"type":"string"},"year":{"type":"number"},"description":{"type":"string"},"website":{"type":"string"},"url":{"type":"string"},"contact_email":{"type":"string"},"hometown":{"type":"string"}},"required":["name","year","description","website","url","contact_email","hometown"]},"CampEntity":{"type":"object","properties":{"uid":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"year":{"type":"number"},"name":{"type":"string"},"url":{"type":"string","nullable":true},"contact_email":{"type":"string","nullable":true},"hometown":{"type":"string","nullable":true},"description":{"type":"string","nullable":true}},"required":["uid","createdAt","updatedAt","year","name","url","contact_email","hometown","description"]},"LocationDto":{"type":"object","properties":{"string":{"type":"string"},"frontage":{"type":"string"},"intersection":{"type":"string"},"intersection_type":{"type":"string"},"dimensions":{"type":"string"},"hour":{"type":"number"},"minute":{"type":"number"},"distance":{"type":"number"},"category":{"type":"string"},"gps_latitude":{"type":"number"},"gps_longitude":{"type":"number"}},"required":["string","frontage","intersection","intersection_type","dimensions","hour","minute","distance","category","gps_latitude","gps_longitude"]},"ArtWithLocationsEntity":{"type":"object","properties":{"uid":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"year":{"type":"number"},"name":{"type":"string"},"url":{"type":"string","nullable":true},"contact_email":{"type":"string","nullable":true},"hometown":{"type":"string","nullable":true},"description":{"type":"string","nullable":true},"artist":{"type":"string","nullable":true},"category":{"type":"string","nullable":true},"program":{"type":"string","nullable":true},"donation_link":{"type":"string","nullable":true},"guided_tours":{"type":"number","nullable":true},"self_guided_tour_map":{"type":"number","nullable":true},"thumbnail_url":{"type":"string","nullable":true},"locations":{"type":"array","items":{"type":"object"}}},"required":["uid","createdAt","updatedAt","year","name","url","contact_email","hometown","description","artist","category","program","donation_link","guided_tours","self_guided_tour_map","thumbnail_url","locations"]},"ArtDto":{"type":"object","properties":{"uid":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"name":{"type":"string"},"year":{"type":"number"},"description":{"type":"string"},"artist":{"type":"string"},"category":{"type":"string"},"program":{"type":"string"},"donation_link":{"type":"string"},"guided_tours":{"type":"number"},"self_guided_tour_map":{"type":"number"},"website":{"type":"string"},"url":{"type":"string"},"contact_email":{"type":"string"},"hometown":{"type":"string"},"locationId":{"type":"number"},"thumbnail_url":{"type":"string"}},"required":["uid","createdAt","updatedAt","name","year"]},"ArtEntity":{"type":"object","properties":{"uid":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"year":{"type":"number"},"name":{"type":"string"},"url":{"type":"string","nullable":true},"contact_email":{"type":"string","nullable":true},"hometown":{"type":"string","nullable":true},"description":{"type":"string","nullable":true},"artist":{"type":"string","nullable":true},"category":{"type":"string","nullable":true},"program":{"type":"string","nullable":true},"donation_link":{"type":"string","nullable":true},"guided_tours":{"type":"number","nullable":true},"self_guided_tour_map":{"type":"number","nullable":true},"thumbnail_url":{"type":"string","nullable":true}},"required":["uid","createdAt","updatedAt","year","name","url","contact_email","hometown","description","artist","category","program","donation_link","guided_tours","self_guided_tour_map","thumbnail_url"]}}}}